+++ test/special-tag-timeout/test.mjs
@@ -1,0 +1,7 @@
+import { test, TestDir, execPlaywrightTest } from '../_helpers/index.mjs';
+
+const testDir = new TestDir(import.meta);
+
+test(testDir.name, () => {
+  execPlaywrightTest(testDir.name);
+});