+++ test/special-tags/test.mjs
@@ -8,8 +8,6 @@ test(testDir.name, () => {
   checkSkipScenario();
   checkSkipFeature();
   checkFailTag();
-  checkTimeoutTag();
-  checkSlowTag();
   checkRetriesTag();
   checkModeTag();
 });
@@ -53,22 +51,6 @@ function checkRetriesTag() {
     `${' '.repeat(2)}test.describe(() => {`,
     `${' '.repeat(4)}test.describe.configure({"retries":2});`,
     `${' '.repeat(4)}test.describe.configure({"retries":1});`,
-  ]);
-}
-
-function checkTimeoutTag() {
-  testDir.expectFileContains(`.features-gen/timeout.feature.spec.js`, [
-    `${' '.repeat(2)}test.describe.configure({"timeout":5000});`,
-    `${' '.repeat(4)}test.setTimeout(4000);`,
-    `${' '.repeat(4)}test.describe.configure({"timeout":3000});`,
-    `${' '.repeat(6)}test.setTimeout(2000);`,
-  ]);
-}
-
-function checkSlowTag() {
-  testDir.expectFileContains(`.features-gen/slow.feature.spec.js`, [
-    `${' '.repeat(2)}test.slow();`,
-    `${' '.repeat(4)}test.slow();`,
   ]);
 }